Introduce Base_with_time_stamp
