+++ Triangulation_2/test/Triangulation_2/test_cdt_degenerate_case.cpp
@@ -2,41 +2,14 @@
 #include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
 #include <CGAL/Constrained_Delaunay_triangulation_2.h>
 #include <CGAL/Constrained_triangulation_plus_2.h>
+#include <CGAL/Base_with_time_stamp.h>
 #include <iostream>
 
 typedef CGAL::Exact_predicates_inexact_constructions_kernel EPIC;
 typedef EPIC::Point_2 Point_2;
 
-template <typename Vb>
-class My_vertex_base : public Vb {
-  std::size_t time_stamp_;
-public:
-  My_vertex_base() : Vb(), time_stamp_(-1) {
-  }
-
-  My_vertex_base(const My_vertex_base& other) :
-    Vb(other),
-    time_stamp_(other.time_stamp_)
-  {}
-
-  typedef CGAL::Tag_true Has_timestamp;
-
-  std::size_t time_stamp() const {
-    return time_stamp_;
-  }
-  void set_time_stamp(const std::size_t& ts) {
-    time_stamp_ = ts;
-  }
-
-  template < class TDS >
-  struct Rebind_TDS {
-    typedef typename Vb::template Rebind_TDS<TDS>::Other Vb2;
-    typedef My_vertex_base<Vb2> Other;
-  };
-};
-
 #ifdef CGAL_CDT_2_DEBUG_INTERSECTIONS
-typedef My_vertex_base<CGAL::Triangulation_vertex_base_2<EPIC> > Vb;
+typedef CGAL::Base_with_time_stamp<CGAL::Triangulation_vertex_base_2<EPIC> > Vb;
 #else
 typedef CGAL::Triangulation_vertex_base_2<K> Vb;
 #endif
